rename to vaultcraft and remove old stuff
