+++ projects/vaultcraft/constants.js
@@ -1,0 +1,21 @@
+const ADDRESSES_ = {
+  ethereum: {
+    vaultRegistry: '0x007318Dc89B314b47609C684260CfbfbcD412864'
+  },
+  arbitrum: {
+    vaultRegistry: '0xB205e94D402742B919E851892f7d515592a7A6cC'
+  },
+  polygon: {
+    vaultRegistry: '0x2246c4c469735bCE95C120939b0C078EC37A08D0'
+  },
+  bsc: {
+    vaultRegistry: '0x25172C73958064f9ABc757ffc63EB859D7dc2219'
+  },
+  optimism: {
+    vaultRegistry: '0xdD0d135b5b52B7EDd90a83d4A4112C55a1A6D23A'
+  },
+}
+
+module.exports = {
+  ADDRESSES: ADDRESSES_
+}
